+++ src/code0452.rs
@@ -60,7 +60,7 @@ impl Solution {
                 endpoint = point[1]
             }
             count
-        }) as i32
+        })
     }
 }
